+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/NussknackerApp.scala
@@ -6,10 +6,11 @@ import pl.touk.nussknacker.ui.factory.NussknackerAppFactory
 object NussknackerApp extends IOApp {
 
   override def run(args: List[String]): IO[ExitCode] = {
-    for {
-      appFactory <- IO(new NussknackerAppFactory(getClass.getClassLoader))
-      _          <- appFactory.createApp().use { _ => IO.never }
-    } yield ExitCode.Success
+    val program = for {
+      appFactory <- NussknackerAppFactory.create(getClass.getClassLoader)
+      app        <- appFactory.createApp()
+    } yield app
+    program.useForever.as(ExitCode.Success)
   }
 
 }