+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/process/processingtype/ScenarioParametersServiceTest.scala
@@ -16,7 +16,10 @@ import pl.touk.nussknacker.restmodel.scenariodetails.ScenarioParameters
 import pl.touk.nussknacker.security.Permission
 import pl.touk.nussknacker.test.ValidatedValuesDetailedMessage
 import pl.touk.nussknacker.test.utils.domain.TestFactory
-import pl.touk.nussknacker.ui.process.processingtype.loader.ProcessingTypesConfigBasedProcessingTypeDataLoader
+import pl.touk.nussknacker.ui.process.processingtype.loader.{
+  DeploymentManagersClassLoader,
+  ProcessingTypesConfigBasedProcessingTypeDataLoader
+}
 import pl.touk.nussknacker.ui.security.api.{LoggedUser, RealLoggedUser}
 import cats.effect.unsafe.implicits.global
 import pl.touk.nussknacker.engine.definition.component.Components.ComponentDefinitionExtractionMode
@@ -284,7 +287,8 @@ class ScenarioParametersServiceTest
     val processingTypeDataReader = new ProcessingTypesConfigBasedProcessingTypeDataLoader(
       new LoadableConfigBasedNussknackerConfig(IO.pure {
         ConfigWithUnresolvedVersion(ConfigFactory.parseFile(devApplicationConfFile).withFallback(fallbackConfig))
-      })
+      }),
+      DeploymentManagersClassLoader.create(List.empty).allocated.map(_._1).unsafeRunSync()
     )
 
     val processingTypeData = processingTypeDataReader