+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/DeploymentManagersClassLoader.scala
@@ -1,0 +1,1 @@
+package pl.touk.nussknacker.engine