+++ utils/utils-internal/src/main/scala/pl/touk/nussknacker/engine/util/config/ConfigWithUnresolvedVersionExt.scala
@@ -24,6 +24,14 @@ class ConfigWithUnresolvedVersionExt(val config: ConfigWithUnresolvedVersion) {
     }
   }
 
+  def readStringList(path: String): Option[List[String]] = {
+    if (config.resolved.hasPath(path)) {
+      Some(config.resolved.getStringList(path).asScala.toList)
+    } else {
+      None
+    }
+  }
+
 }
 
 object ConfigWithUnresolvedVersionExt {